refactor: reorganize module structure
- Move SoapClient trait, SoapResponse, encode_envelope to soap_client/mod.rs
- Convert decode_fault to SoapResponse::fault_message() method
- Remove duplicated code from lib.rs and local.rs
- Update imports throughout codebase
